\name{FLIM-internal} 
\alias{ExpandDataset} 
\alias{FillInGaps} 
\alias{ImputateGaps}
\alias{FlimFormula}  
\alias{ImputateMissingValues}
\alias{SetImputationTimes}
\alias{SetIncrements}
\alias{SetStaticVariables}
\alias{spaplot} 
\title{Internal mypackage Functions}
\description{Internal \code{FLIM} functions} 
\details{These are not to be called by the user.} 
\keyword{ internal } 